#194 Ability to decorate task execution
