+++ db-scheduler/src/main/java/com/github/kagkarlsson/scheduler/task/Task.java
@@ -15,48 +15,16 @@
  */
 package com.github.kagkarlsson.scheduler.task;
 
-public abstract class Task<T> implements ExecutionHandler<T> {
-    protected final String name;
-    private final FailureHandler<T> failureHandler;
-    private final DeadExecutionHandler<T> deadExecutionHandler;
-    private final Class<T> dataClass;
+public interface Task<T> extends ExecutionHandler<T> {
+    String getName();
 
-    public Task(String name, Class<T> dataClass, FailureHandler<T> failureHandler, DeadExecutionHandler<T> deadExecutionHandler) {
-        this.name = name;
-        this.dataClass = dataClass;
-        this.failureHandler = failureHandler;
-        this.deadExecutionHandler = deadExecutionHandler;
-    }
+    Class<T> getDataClass();
 
-    public String getName() {
-        return name;
-    }
+    TaskInstance<T> instance(String id);
 
-    public Class<T> getDataClass() {
-        return dataClass;
-    }
+    TaskInstance<T> instance(String id, T data);
 
-    public TaskInstance<T> instance(String id) {
-        return new TaskInstance<>(this.name, id);
-    }
+    FailureHandler<T> getFailureHandler();
 
-    public TaskInstance<T> instance(String id, T data) {
-        return new TaskInstance<>(this.name, id, data);
-    }
-
-    public abstract CompletionHandler<T> execute(TaskInstance<T> taskInstance, ExecutionContext executionContext);
-
-    public FailureHandler<T> getFailureHandler() {
-        return failureHandler;
-    }
-
-    public DeadExecutionHandler<T> getDeadExecutionHandler() {
-        return deadExecutionHandler;
-    }
-
-    @Override
-    public String toString() {
-        return "Task " + "name=" + getName();
-    }
-
+    DeadExecutionHandler<T> getDeadExecutionHandler();
 }
